Fix old ratelimit resets breaking next requests
If the reset was in the past by a second, then multiplication to
determine the difference in time to wait would result in somewhere
around the max value of u64 being multiplied by 1000, which caused a
multiplication overflow.

Instead, if the retry is in the past, just ignore it, as it would have
already passed, and continue on with the request, setting the remaining
requests allowed by the bucket to the known limit.