Framework: fix command arg positioning
The command system assumed that prefixes were only one character long,
so count the total length of the prefix. In addition, the
`allow_whitespace` configuration added some difficulty in deciding where
to count as the initial position to start splitting for arguments.

Instead of fixing that, rewrite the framework to make these types of
changes easier in the future.