Nonblocking connection
The connection is now non-blocking, allowing user event handlers to
immediately unlock it themselves (assuming they haven't unlocked it
elsewhere), rather than waiting on the library to receive an event.

This is done by decoupling the receiver away from the connection,
which is not necessarily "directly" associated with the connection.

This needs a _lot_ of code cleanup later.