Change the CACHE to be an RwLock
The global Cache used to be an Arc<Mutex>, however the issue is that it
could only be opened for reading or writing once at a time.

With an RwLock, multiple readers can access the Cache at once, while
only one Writer may at once. This will allow users to be able to have
multiple Readers open at once, which should ease some of the pains with
working with the Cache.

Upgrade path:

Modify all uses of the CACHE from:
`CACHE.lock().unwrap()`
to
`CACHE.read().unwrap()` if reading from the Cache (most use cases), or
`CACHE.write().unwrap()` to write to it.