Use html_logo_url instead of a doc script
The script was made with the eventual purpose to build custom docs.
This has been replaced by [cargo-external-doc], which was released a
few days after the script was added.

Since the only other purpose of the script besides adding the image has
been fulfilled by another crate, we can just utilize a rustdoc attribute
to do this.