Fix unreachable patterns in command macro
When attempting to use a named argument with a type of String, the
framework will try to parse the received argument - a String - into a
String. This can never fail, so Rust warns that the `Err` arm of the
`match` can never be reached.

Let's just ignore that and everything will be fine.