Check for embed lengths on message sends
Due to the maximum length of the textual content of an embed being 4000,
this should be checked prior to message sending.

Embeds have a fairly new limit of only being 4000 character long, at
maximum. The length of these fields - where present - should be summed,
so that bad requests are not made with embeds that are too large in
text.

The fields that count as "textual" includes:

- author name
- description
- `name` and `value` of each field
- footer text
- title

If this exceeds the limit - currently at 4000 unicode codepoints - then
a `ClientError::EmbedTooLarge` will be returned, including the number
of codepoints overflowing.