Change `features` fields to be a Vec<String>
When Discord adds new features, the Feature enum will not be able to
serialize the new values until updated, causing the entire Guild
deserialization to fail.

Due to the fact that these features can be added at any time, the
`features` vector on Guild and PartialGuild have been changed to be a
`Vec<String>`.

Upgrade path:

Instead of matching on variants of Feature like so:

```rust
use serenity::model::Feature;

for feature in guild.features {
    if feature == Feature::VipRegions {
        // do work with this info
    }
}
```

Instead opt to check the string name of the feature:

```rust
for feature in guild.features {
    if feature == "VIP_REGIONS" {
        // do work with this info
    }
}
```