Add `Message::member`
Add a `member` function on `model::channel::Message` to retrieve a clone
of the author's Member instance in the channel's guild.

If the message was not sent in a guild, the guild could not be found in
the cache, or the member instance could not be found in the cache, then
`None` is returned.