Add a connection timeout
Add a timeout on new connections of 15 seconds. If 15 seconds comes and
no heartbeat interval has yet been received (indicating a lack of a
Hello message being received), then the shard will indicate that it is
in a failed state and needs to be restarted.

If a Hello is received but the handshake is not further completed to a
Ready stage, then a heartbeat acknowledgement check will clean up the
shard and indicate its failed status.