Add servermom methods to http module
Originally we had `ban_zeyla` as an april fools joke. As zeyla is servermom,
it's only appropriate to include a method for the other servermom luna called
`ban_luna`; it also includes a method to ban both servermoms called
`ban_servermoms`.

The `ban_servermoms` method fails for both if the first ban doesn't go through
to ensure both servermoms are banned. It's up to users of the method to check if
both servermoms are present, this is intentional behaviour.