Don't delay Ready with cache enabled
When the cache is enabled, don't delay the Ready until all guilds have been
received.

This never really worked in the first place and duplicates the "cached" logic
that fires when all guilds have been received.

This presumably fixes the "silent death" bug, as this appears to stall the
thread on certain conditions.