Change default branch to 'current'
The default branch of 'master' has a few issues, but the biggest technical issue
is that it doesn't accurately represent the development work done on it by name
alone: is it the branch for maintaining the "current" released version, or for
the development of the next majour version? In certain Rust projects these are
combined into one branch, but we don't do this, so this is one way out of the
ambiguity.