prelude: re-export typemap::sharemap
When the `typemap` feature is enabled (thus enabling the crate),
re-export the library's `ShareMap` type from serenity's `prelude`
module.

This is a useful re-export due to its use in the Client, which almost
every serenity user will be using. This avoids a direct dependency in
the user's Cargo.toml.

Examples have been updated to use this prelude instead of depending on
the `typemap` crate directly.

Closes #469.