Replace `Context` as Argument (#606)
Prior this commit, if a function needed the HTTP-client and optionally was able to utilise the cache too, it expected `Context`. This causes issues when such a function was called outside the framework.

This commit adds the `CacheHttp`-trait and implements it for `Context`, `Http`, and a tuple of `(CacheRwLock, Http)`.
`Context`s has been replaced with `impl CacheHttp` and therefore accepting any of the mentioned above. If the passed type does not contain a cache, the function will behave as if no `cache`-feature is enabled.

If a function requires HTTP-client and cache, it will expect a generic-bound of `AsRef<CacheRwLock> + AsRef<Http>`, which is also implemented for the tuple of `(CacheRwLock, Http)` if the `cache`- and `http`-feature are enabled.