http/client: store inner client in an arc
In the 'http' module's 'Http' struct, store the inner reqwest client in
an Arc. This doesn't have much use right now for serenity itself, but it
does allow users to pass in a reference to their own client when making
a client via 'http::Http::new', avoiding the need to have multiple
clients.

The primary use for this is to ease a slight rework of the
'http::ratelimiting' module.

Signed-off-by: Zeyla Hellyer <zeyla@hellyer.dev>