ratelimiting: simplify parse_headers, add tests
In the 'http::ratelimiting' module, simplify 'parse_headers' so that
it's no longer a chain of Result combinators. This has no functional
difference, but is simply for ease of readability.

Additionally, add some tests for 'parse_headers'. These tests were added
before the simplification of the function and tested, and then used to
test the function rewrite.

Signed-off-by: Zeyla Hellyer <zeyla@hellyer.dev>