model/guild: fix 'after' generic on 'members'
In the 'Guild', 'GuildId', and 'PartialGuild' structs, there is a method
named 'members' to retrieve a list of members from the REST API with
some parameters.

One of these parameters was to retrieve a subsection of members after a
certain ID, optionally passed.

The parameter type in universal trait form looked like this:

```rust
after: Option<impl Into<UserId>>
```

If someone passes 'Some(u64)', then the inner value is mapped with its
'From u64 for UserId' impl. However, if someone passes a 'None', then we
don't actually know what the type of the Option is -- T is not defined.
The user needs to do something like 'None::<UserId>', 'None::<u64>',
etc.

This fixes that discrepency, by allowing users to pass anything that
converts into an 'Option<UserId>'. Users can pass in 'UserId(1)',
'None', etc. now, without the weirdness around the generic within the
Option.

Signed-off-by: Zeyla Hellyer <zeyla@hellyer.dev>