http: rework ratelimiting structure
Rework the ratelimiting structure to avoid an odd hierarchy and to
remove the static variable.

Ratelimiting was done through a freefloating function named 'perform'.
It accepted a reference to the Http client instance and the request to
perform. This is backwards in design: the Http client in this design is
passing itself to the ratelimiting module, which in turn calls back into
the Http passed into it to access route ratelimit information and the
global ratelimiter mutex on the Http client. At that point, you may as
well put the ratelimiting on the Http client itself.

Another poor design choice was to use a 'static mut' for storing the
offset. Although this is currently only set once -- after the first
response ever is received -- it's noted[1] in the source right now that
it may be preferable to mutate it more often in the future.

The first issue fixed is the ratelimiting design: the ratelimiter module
now has a struct named 'ratelimiter' that owns:

- its own global ratelimiting mutex
- an Arc<ReqwestClient> that it shares with the Http client
- the offset stored as an atomic
- a map of routes to their ratelimits
- the bot token

With this design, ratelimiting information is self-contained.

The second issue was the offset. The public interface is the same here,
but now on the 'ratelimiter' struct. That is, the function definition
still returns an 'option<i64>', despite the fact that it's internally
stored as an atomic i64. The value is i64::MAX when no offset has been
calculated yet.

While not perfect still, there have been some efficiency improvements.
The lock for each route is no longer held throughout the duration of
each request to that route. This means that 5 message sends in the same
channel, assuming a limit of 5, will happen more or less simultaneously
if sent all in one go on 5 threads.

A minour breaking change is that the 'ratelimit' struct containing the
'limit', 'remaining', and 'reset' for a route are now private
structfields, exposed through public getters. This probably doesn't
affect any users.

Signed-off-by: Zeyla Hellyer <zeyla@hellyer.dev>