model/guild: fix guild deserialization regression
A regression in 'model::guild::Guild' deserialization was introduced by
commit 4541935243d794cc760f46520f0cfa3f4994a9a3[1].

The bug introduced was that deserialization of
'Guild::premium_subscription_count'[2] would fail. The
'premium_subscription_count' field isn't always present in the JSON
provided as noted by the REST API documentation[3], so a default value
needs to be set for it, in this case 0.

Additionally, set a default value for 'Guild::premium_tier'[4] in case this
is ever nulled in the API too, defaulting to 'PremiumTier::Tier0'.

[1]: https://github.com/serenity-rs/serenity/commit/4541935243d794cc760f46520f0cfa3f4994a9a3
[2]: https://github.com/serenity-rs/serenity/commit/4541935243d794cc760f46520f0cfa3f4994a9a3#diff-e1c86b244465a2bb338d53758aa2c7efR154
[3]: https://github.com/discordapp/discord-api-docs/commit/3d8cca0c3c5cf0a42458c9d7c6137534d12a46bc#diff-c4e1390fca89ecc96b01a093f988fa1cR48
[4]: https://github.com/serenity-rs/serenity/commit/4541935243d794cc760f46520f0cfa3f4994a9a3#diff-e1c86b244465a2bb338d53758aa2c7efR156

Signed-off-by: Zeyla Hellyer <zeyla@hellyer.dev>