Get rid of the `Arc` implementation for `Framework`
This is technically a breaking change, but I doubt its removal will affect anyone.
There's no meaningful usage of an `Arc<T> where T: Framework` without `Mutex` or `RwLock`-based interior-mutability indirection
as the framework method requires mutable access of `self` (`&mut self`).
This can be hard to achieve with `Arc` because there must be only one owner of the pointer at the time, nullifying its usefulness for shared memory.

In summary, this impl is useless.