Update audit log model to reflect Discord's optional/mixed types
There is an error in the currently used data structure serenity which causes Json errors on models where serenity does not expect null values, or mixed type values. While this particular change does not properly wrap all the types, it instead focuses on ensuring it works, not that it works well.

Good enough for a temporary solution I'd say.