First draft of time-based events for audio.
Messy and likely inefficient at this stage, but provides a working example
of timer events tied to the lifetime of an audio track. What sets these
"local" events apart from an external event system is that these
depend on (and fire using) the actual track position, aware of
play status and so on without intrusively locking or inspecting
the audio thread's state.