Reduce full-buffer copies in audio mixing.
Before, every audio source was effectively copied twice:

* Once, into a local buffer,
* Then into the shared floating point buffer.

This changes how `AudioSource`s work so that audio data is immediately
converted into floating point and placed into the shared buffer,
requiring fewer accesses and fewer copies. We've gone from 2 x 3ms per
source to 2.5ms per source with the below improvements.

Additionally, this now places `BufReader`s in front of all piped
audio sources to reduce some of the overhead of requesting pipe
access repeatedly; currently I've set it to buffer up to 1s of frames.
We now convert ffmpeg output to floating point so that we don't have to
do that work ourselves. In spite of the doubled bandwidth over the
pipes, these

There's still an issue with ffmpeg + ytdl, where everything is locked
to 1x speed (so it takes 20ms to read in a frame of audio from each
source, noting that we have a strict 20ms budget for ALL sources). This
isn't present if only ffmpeg is used, and ytdl on its own will download
the file very quickly.

(This still includes some instrumentation printouts...)