First draft of a basic seekable source
`RestartableSource`s allow for cheap forward seeking, and restart the
download + decompression chain whenever a backwards seek is required.
This should be usable with ffmpeg right now, but not ytdl -- the
mechanism works, most importantly.

I've laid out some other sources I want to include, but I have a feeling
that to put them in cleanly I'm as well replacing the `AudioSource`
trait with an `impl` of `Read` to simplify the design (and then the
top-level methods like `add_float_pcm_frame` would apply to any `Read` generic).