Machinery, sketch of implementation for memory buffers
I have this worked out in my head, but it's too late to write the code and I'd rather not lose progress.

Unsafe use is mainly because this relies upon some trickiness to be thread-safe (any shared handle could write at any time, and locking around most of the data is disastrous). I intend to write a blog post or something to explain this, that way implementation and design can be cross-checked against one another.