Memory-buffered sources
Implements read and seek capability for fully buffered sources. These allow quick seeking, quick reading, and shared (almost-)lockless access to one-off streams (like ffmpeg) between many threads at the cost of RAM. The intended use case is small, constantly used/shared sounds (i.e., sound effects, stings played every time a song is queued, so on).

Essentially, a rope of chunks is maintained, and any reads transparently write chunks into this shared struct. No handle attempts to lock this struct unless:
 * A read goes past the end of the currently saved bytes, and
 * The store is not yet finalised (i.e., read Ok(0) from the input).
Because handling the critical section in this way is very hard to express in safe rust, we need to employ unsafe code to have shared mutability. On finalisation, bytes are copied form the rope to a contiguous store, leading to very fast reads. In lieu of metadata, length hints can be used to estimate the required space (and reduce runtime/finalisation cost).

This has some drawbacks for now: a poor choice of chunk length will cause mixing times to increase due to linked list pointer chasing (note that the use of an actual list is key to thread-safety: no resizing). Designing a thread-safe lockless acceleration structure would be interesting, but risky. I was looking into storing pointers to the last accessed member of the rope, but I considered this a bridge too far until correctness could be guaranteed. I hope to include some options to spawn a thread to take care of finalisation: right now, this has the potential to stall the audio thread if a poor length hint / chunk length is provided.

I intend to write this up more clearly somewhere, so that the algorithm and implementation can be separated neatly. Next up should be using this mechanism to store Opus frames representing the input data.