Looping support, event handling, and restartable ytdls
Added some examples on how loop setting works as an extra sound effect in example 15. Tracks can freely loop now, just as long as they are `Seek`able -- this can be a finite or infinite count! This should now fire a new event class (again, made us of in example 15).