Redesign client to use a builder impl `Future`.
This design allows to create a client builder, chain methods on it,
and then await the builder, returning a client if successful.

Additionally, one can provide an `Arc`'ed framework and keep an
`Arc` now.
This allows to keep the framework for manual dispatch.

Last but not least, event handlers are no longer required for
the framework.