Cleanup of `cycle`.
Bitrate does not need to be set every time, so we no longer do this.
A few buffers existed due to the old implementation, which are now
removed. Equivalently, opus frame passthrough should now fit into a
constant-size block of stack memory (if and when it is implemented).
The possibility of doing so was causing a Vec allocation/malloc
every frame.