voice: use `xsalsa20poly1305` crate for encryption
This commit replaces the `sodiumoxide` crate, a wrapper for the
libsodium C library, with the pure Rust `xsalsa20poly1305` crate from
the RustCrypto project:

https://github.com/RustCrypto/AEADs/tree/master/xsalsa20poly1305

This should reduce compile times as `libsodium` is a somewhat large
C library and you are only using one algorithm from it. The
`xsalsa20poly1305` crate, on the other hand, is a lightweight pure Rust
dependency that implements only the algorithm you are using.

Note that the advertised MSRV of this crate is 1.41+, however earlier
today new release of the dependency causing that, `generic-array`,
reduced the MSRV to 1.36.0:

https://github.com/fizyk20/generic-array/pull/102

You can read more about RustCrypto and our recent releases in this post:

https://www.reddit.com/r/rust/comments/h0em5n/ann_new_rustcrypto_releases_aead_blockcipher/