Replace the in-house typemap with typemap_rev
I migrated the serenity typemap over to a standalone crate. Doing this
removes an extra file (typemap.rs) which is now unnecessary. I am planning
on publishing the crate on crates.io in the not-too-distant future, but for now,
the git is the same as the current typemap utility.

I have invited acdenis to the typemap_rev repo so he can modify it as needed
since it is mostly his code.

Signed-off-by: kingbri <bdashore3@gmail.com>