Fix shutdown regression
In something that was fixed before
(https://github.com/serenity-rs/serenity/pull/713), there was an
implementation in ShardManager's shutdown method that forced it to wait
for a response of the ShardRunner that a CloseFrame was sent and
acknowledged.

Somewhere inbetween then and now someone moved the waiting code to a
different thread, thus making the shutdown method non-blocking again and
regressing on the error.

This PR seems to rectify that by;
1. Moving the blocking code back to shutdown.
2. Actually .awaiting sending the CloseFrame