Make parse_emoji use animated emojis
Previously, parse_emoji only returned true when the emoji was static
and animated emojis passed to the function returned None since they
include an extra character (<a:) to denote whether the emoji is animated.

Since we can tell whether an emoji is animated through the mention,
add support in the parse_emoji function to detect if an emoji is animated
or not.

Also alter the EmojiIdentifier struct to include the animated bool.

Signed-off-by: kingbri <bdashore3@gmail.com>