Fix Tungstenite IO error being discarded in ws_impl (#975)
This commit fixes the issue where the `recv_json()` impl in ws_impl was
discarding tungstenite errors, so that any matches later in the error
propagation chain can properly handle this event.

This commit also removes the (imho, nonsensical) IO error discarding
until the next heartbeat fails. An IO error on a connection is, in most
cases, unrecoverable. So it's often faster and more reliable to just
discard the connection and retry.