The great (voice) crate escape.
At last, voice is moved out into its own crate. It's horribly messy, not all feature combinations play nicely, and we don't have cross-library compat yet.

Serenity now should support any compatible voice driver, using a trait to handle gateway responses to voice events and provide a handle for each shard's intermessages. This mimics how frameworks and eventhandlers are already implemented.

One of the big user-facing implications of this is that installing the voice client is now two lines at most: a trait import, and a call on the `ClientBuilder`. This is a huge step up ergonomically from having to define a TypeMapKey, manually insert the voice driver after taking it from the client, etc. It still works like this internally, but access to the voice manager is now as simple as calling `songbird::get(ctx)`.

Unanswered questions:
* What happens to the parts in `model::*` which only appear in the voice gateway API? Do we move these into the subcrate? I don't believe twilight handle these events at all, for instance.
* Do we still need the "voice" feature in the core serenity crate? This now gates access only to the voice client trait, and adds Option<...> fields to clients, runners and other managers.

Unrelated:
* Identified (probably) how to fix rebalance voice loss bug. Fixing this will probably come alongside a minor redesign of the management system.

Cleanup ahead, but it's one of the larger tasks I was concerned about.
* Issue has cropped up in example 14 with some sources.