Manager redesign
Removes one layer of indirection, and should fix the voice rebalance bug. Not had any occur to test, however.

The main change is that when using Serenity as the gateway backend, rebalances should no longer completely remove the individual voice managers. Rather, the channel to send messages back to each shard runner is stored in a shared location -- when a shard is forced to disconnect, senders will start queueing up messages to be transmitted. On completion, the new channel is inserted and buffered messages are cleared.

TODO: look into dashmap for the main call store? I couldn't figure out a clean "insert if not present" syntax.

Works for e06. Rest of the examples need to be ported, however.