Handle Voice close codes, prevent Songbird spinning WS threads.
Voice CloseCodes now map to a type rather than a collection of constants. Correct close code handling in this way terminates the ws task when there is no likelihood of resuming, which was causing leftover tasks to spin at the tokio::select in some circumstances (i.e., ::leave, which keeps the Driver alive).