Check a sub-command's parent first.
Issue:
When the user of the help-system issues a request for a sub-command, the help-system would ignore the parent's configuration.
Hence the sub-command was executed without the parent's consideration.

Example:
A parent command is limited to guild channels only and defines a child.
The child command has no configuration.
The user may not invoke help for the parent command, but for its child.

Fix:
The help-system runs a `command_behaviour`-check on the parent first, before picking a matching sub-command.