Bump to 0.10.2 (#1182)
This release contains a compatibility patch for tokio, as much of the ecosystem is still catching up in regards to the `1.0` release. The patch will allow users to use tokio `0.2` until all of their dependencies are updated to use `1.0`. This release also fixes documentation attributes for slash commands, correctly displaying information that their api is unstable and has to be enabled with the `unstable_discord_api` feature.