Preserve documentation comments for the help command (#1232)
This fixes a compilation error if a user provides documentation comments to the `#[help]` macro:
```
invalid attribute: "description"
```

The help macro uses similar code to parse the function it is applied on, which converts `#[doc = "..."]` attributes (desugared form of `///` comments) to `#[description = "..."]`. This commit reverses that action to allow usage of `///` comments on help functions.